Add files via upload
Title of our project is “Hotel Management System”. We have tried to show how the data/information in hotels is managed. This is just an overview of management in hotels. It manages and maintains the records of customers, rooms, employees and drivers in the hotel. The project is aimed to maintain the day-to-day state of admission /vacation of Residents, List of employees, room details etc. Main objective of this project is to provide a solution for hotels to manage most their work using computerized processes.This software application will help admin to handle customer information, room allocation details, Payment details etc.The rooms have different categories like single bed, double bed etc. so their charges and records will be maintained accordingly.This software has been made in a user friendly interface, so that anyone can add, delete, update the entries and handle all the transactions easily. As a security I have provided Admin username and Password.
The project, Hotel Management System is a desktop-based application that allows the hotel manager to handle all hotel activities online. Interactive GUI and the ability to manage various rooms, employees, drivers and customers make this system very flexible and convenient. This application gives managers the power and flexibility to manage the entire system from a single online system. Hotel management project provides room booking, staff management and other necessary hotel management features. The system allows the manager to post available rooms in the system.